src/api/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
mysql-connector